// interfaces/users.ts
export interface UserProfile {
    id?: string;
    firstName?: string;
    lastName?: string;
    userName?: string;
    email: string;
    phoneNumber?: string;
    ssn?: string;
    registered_at?: Date;
    profile?: string;
    password?: string;
    idNumber?: string;
    last_login?: Date;
    status?: number;
    role_id?: number;
}